+++ backend/src/main/java/com/pickeat/backend/room/application/dto/RoomService.java
@@ -1,4 +1,4 @@
-package com.pickeat.backend.room.application;
+package com.pickeat.backend.room.application.dto;
 
 import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.global.exception.ErrorCode;
@@ -10,9 +10,6 @@ import com.pickeat.backend.room.domain.RoomUser;
 import com.pickeat.backend.room.domain.repository.RoomRepository;
 import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.room.domain.repository.RoomUserRepository.RoomUserCount;
-import com.pickeat.backend.user.domain.repository.UserRepository;
-import com.pickeat.backend.wish.domain.WishList;
-import com.pickeat.backend.wish.domain.repository.WishListRepository;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -21,32 +18,28 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
+@Service("RoomServiceV2")
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class RoomService {
 
+    private static final int INITIAL_ROOM_COUNT = 1;
     private final RoomRepository roomRepository;
-    private final UserRepository userRepository;
     private final RoomUserRepository roomUserRepository;
-    private final WishListRepository wishListRepository;
 
     @Transactional
     public RoomResponse createRoom(RoomRequest request, Long userId) {
         Room room = new Room(request.name());
         roomRepository.save(room);
         roomUserRepository.save(new RoomUser(room.getId(), userId));
-
-        WishList wishList = createWishList(room);
-        return RoomResponse.of(room, getRoomUserCount(room), wishList.getId());
+        return RoomResponse.of(room, INITIAL_ROOM_COUNT);
     }
 
     public RoomResponse getRoom(Long roomId, Long userId) {
         validateUserAccessToRoom(roomId, userId);
 
         Room room = getRoomById(roomId);
-        long wishListId = getWishListId(roomId);
-        return RoomResponse.of(room, getRoomUserCount(room), wishListId);
+        return RoomResponse.of(room, getRoomUserCount(room));
     }
 
     public List<RoomResponse> getAllRoom(Long userId) {
@@ -63,8 +56,7 @@ public class RoomService {
         return rooms.stream()
                 .map(room -> RoomResponse.of(
                         room,
-                        roomUser.getOrDefault(room.getId(), 0),
-                        getWishListId(room.getId())
+                        roomUser.getOrDefault(room.getId(), 0)
                 ))
                 .toList();
     }
@@ -90,18 +82,6 @@ public class RoomService {
     @Transactional
     public void exitRoom(Long roomId, Long userId) {
         roomUserRepository.deleteByRoomIdAndUserId(roomId, userId);
-    }
-
-    private WishList createWishList(Room room) {
-        WishList wishList = new WishList(room.getName(), room.getId(), false);
-        wishListRepository.save(wishList);
-        return wishList;
-    }
-
-    private long getWishListId(Long roomId) {
-        WishList wishList = wishListRepository.findByRoomIdAndIsTemplate(roomId, false)
-                .orElseThrow(() -> new BusinessException(ErrorCode.WISH_LIST_NOT_FOUND));
-        return wishList.getId();
     }
 
     private int getRoomUserCount(Room room) {
