+++ backend/src/main/java/com/pickeat/backend/room/application/dto/request/RoomRequest.java
@@ -3,7 +3,7 @@ package com.pickeat.backend.room.application.dto.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 
-@Schema(description = "방 생성 요청")
+@Schema(description = "방 생성 요청", name = "RoomRequestV2")
 public record RoomRequest(
 
         @NotBlank(message = "방 이름은 공백일 수 없습니다.")
