+++ backend/src/main/java/com/pickeat/backend/room/application/dto/response/RoomResponse.java
@@ -3,19 +3,17 @@ package com.pickeat.backend.room.application.dto.response;
 import com.pickeat.backend.room.domain.Room;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-@Schema(description = "방 정보 응답")
+@Schema(description = "방 정보 응답", name = "RoomResponseV2")
 public record RoomResponse(
         @Schema(description = "방 ID", example = "1")
         long id,
         @Schema(description = "방 이름", example = "점심 같이 먹어요")
         String name,
         @Schema(description = "방에 참여한 인원 수", example = "3")
-        int userCount,
-        @Schema(description = "방의 위시리스트 ID", example = "1")
-        long wishlistId
+        int userCount
 ) {
 
-    public static RoomResponse of(Room room, int userCount, long wishlistId) {
-        return new RoomResponse(room.getId(), room.getName(), userCount, wishlistId);
+    public static RoomResponse of(Room room, int userCount) {
+        return new RoomResponse(room.getId(), room.getName(), userCount);
     }
 }
