+++ backend/src/main/java/com/pickeat/backend/tobe/wish/application/WishPictureService.java
@@ -4,7 +4,7 @@ import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.global.exception.ErrorCode;
 import com.pickeat.backend.restaurant.domain.Picture;
 import com.pickeat.backend.restaurant.domain.RestaurantInfo;
-import com.pickeat.backend.tobe.room.domain.repository.RoomUserRepository;
+import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.wish.application.dto.request.ImageRequest;
 import com.pickeat.backend.tobe.wish.application.dto.response.WishPictureResponse;
 import com.pickeat.backend.tobe.wish.domain.Wish;
