+++ backend/src/main/java/com/pickeat/backend/tobe/wish/application/WishService.java
@@ -5,8 +5,8 @@ import com.pickeat.backend.global.exception.ErrorCode;
 import com.pickeat.backend.restaurant.domain.FoodCategory;
 import com.pickeat.backend.restaurant.domain.RestaurantInfo;
 import com.pickeat.backend.room.domain.Room;
-import com.pickeat.backend.tobe.room.domain.repository.RoomRepository;
-import com.pickeat.backend.tobe.room.domain.repository.RoomUserRepository;
+import com.pickeat.backend.room.domain.repository.RoomRepository;
+import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.wish.application.dto.request.WishRequest;
 import com.pickeat.backend.tobe.wish.application.dto.request.WishUpdateRequest;
 import com.pickeat.backend.tobe.wish.application.dto.response.WishResponse;
