+++ backend/src/test/java/com/pickeat/backend/tobe/wish/application/WishPictureServiceTest.java
@@ -14,9 +14,9 @@ import com.pickeat.backend.global.exception.ErrorCode;
 import com.pickeat.backend.restaurant.domain.Picture;
 import com.pickeat.backend.room.domain.Room;
 import com.pickeat.backend.room.domain.RoomUser;
+import com.pickeat.backend.room.domain.repository.RoomRepository;
+import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.fixture.WishFixture;
-import com.pickeat.backend.tobe.room.domain.repository.RoomRepository;
-import com.pickeat.backend.tobe.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.user.domain.repository.UserRepository;
 import com.pickeat.backend.tobe.wish.application.dto.response.WishPictureResponse;
 import com.pickeat.backend.tobe.wish.domain.Wish;
